new storage integration: fix some icogrid tests

+++ include/stencil-composition/icosahedral_grids/iterate_domain.hpp
@@ -461,7 +461,7 @@ namespace gridtools {
             // i+offset_i or j+offset_j or k+offset_k is too large.
             // Most probably this is due to you specifying a positive offset which is larger than expected,
             // or maybe you did a mistake when specifying the ranges in the placehoders definition
-            GTASSERT(storage_info->size() > pointer_offset);
+            GTASSERT(static_cast<int>(storage_info->size()) > static_cast<int>(pointer_offset));
             return static_cast< const IterateDomainImpl * >(this)
                 ->template get_value_impl<
                     typename iterate_domain< IterateDomainImpl >::template accessor_return_type< Accessor >::type,
@@ -495,7 +495,7 @@ namespace gridtools {
             const storage_info_t *storage_info =
                 boost::fusion::at< storage_info_index_t >(m_local_domain.m_local_storage_info_ptrs);
 
-            GTASSERT(storage_info->size() > offset);
+            GTASSERT(static_cast<int>(storage_info->size()) > static_cast<int>(offset));
 #endif
             return static_cast< const IterateDomainImpl * >(this)
                 ->template get_value_impl<
